chore: move constant to remove dep
