+++ packages/agent-context/src/insights/classifyConversation.ts
@@ -2,7 +2,7 @@ import type {SanityClient} from '@sanity/client'
 import {generateText, type LanguageModel, Output} from 'ai'
 import {z} from 'zod'
 
-import {CONVERSATION_SCHEMA_TYPE_NAME} from '../studio/insights/schemas/conversationSchema'
+import {CONVERSATION_SCHEMA_TYPE_NAME} from './constants'
 import type {Message} from './saveConversation'
 
 /** @public */
