+++ packages/agent-context/src/insights/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * Schema type name for conversation documents.
+ *
+ * @public
+ */
+export const CONVERSATION_SCHEMA_TYPE_NAME = 'sanity.agentContextConversation'
