+++ packages/agent-context/src/studio/index.ts
@@ -1,10 +1,8 @@
+export {CONVERSATION_SCHEMA_TYPE_NAME} from '../insights/constants'
 export {
   AGENT_CONTEXT_SCHEMA_TYPE_NAME,
   agentContextSchema,
 } from './context-document/agentContextSchema'
-export {
-  CONVERSATION_SCHEMA_TYPE_NAME,
-  conversationSchema,
-} from './insights/schemas/conversationSchema'
+export {conversationSchema} from './insights/schemas/conversationSchema'
 export type {AgentContextPluginOptions, InsightsOptions} from './plugin'
 export {agentContextPlugin} from './plugin'
